+++ engine/src/main/kotlin/star/galaxy/engine/Engine.kt
@@ -3,5 +3,5 @@ package star.galaxy.engine
 import star.galaxy.engine.metainformations.Second
 
 interface Engine {
-    fun nextStep(@Second Δt: Int)
+    fun nextStep(@Second Δt: Double)
 }
